chore: rename to rex-xai throughout
