+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "rex-ai"
+name = "rex-xai"
 version = "0.1.2"
 description = "causal Responsibility-based eXplanations of black-box-classifiers"
 authors = ["David Kelly, <dkellino@gmail.com>"]
@@ -24,7 +24,7 @@ pillow = ">= 10.3.0"
 torch = ">= 2.4.0"
 
 [tool.poetry.scripts]
-ReX = "rex_ai.rex:main"
+ReX = "rex_xai.rex:main"
 
 [build-system]
 requires = ["poetry-core"]
@@ -43,5 +43,5 @@ pyright = "^1.1.383"
 pytest-cov = "^5.0.0"
 
 [tool.pyright]
-include = ["rex_ai"]
+include = ["rex_xai"]
 reportMissingTypeStubs = false
