+++ rex_xai/explanation.py
@@ -11,16 +11,16 @@ import torch as tt
 import numpy as np
 from PIL import Image 
 
-from rex_ai.evaluation import Evaluation
-from rex_ai.extraction import Explanation
-from rex_ai.responsibility import causal_explanation
-from rex_ai.input_data import Data
-from rex_ai.onnx import get_prediction_function
-from rex_ai.resp_maps import ResponsibilityMaps
-from rex_ai.occlusions import set_mask_value
-from rex_ai.config import CausalArgs
-from rex_ai.logger import logger, set_log_level
-from rex_ai.database import initialise_rex_db, update_database
+from rex_xai.evaluation import Evaluation
+from rex_xai.extraction import Explanation
+from rex_xai.responsibility import causal_explanation
+from rex_xai.input_data import Data
+from rex_xai.onnx import get_prediction_function
+from rex_xai.resp_maps import ResponsibilityMaps
+from rex_xai.occlusions import set_mask_value
+from rex_xai.config import CausalArgs
+from rex_xai.logger import logger, set_log_level
+from rex_xai.database import initialise_rex_db, update_database
 
 
 def try_preprocess(args, model_shape, device):
