+++ rex_xai/responsibility.py
@@ -16,13 +16,13 @@ except ImportError:
     from anytree.search import find
 
 
-from rex_ai.box import average_box_size, initialise_tree
-from rex_ai.config import CausalArgs, Queue
-from rex_ai.input_data import Data
-from rex_ai.logger import logger
-from rex_ai.mutant import Mutant, get_combinations
-from rex_ai.resp_maps import ResponsibilityMaps
-from rex_ai.prediction import Prediction
+from rex_xai.box import average_box_size, initialise_tree
+from rex_xai.config import CausalArgs, Queue
+from rex_xai.input_data import Data
+from rex_xai.logger import logger
+from rex_xai.mutant import Mutant, get_combinations
+from rex_xai.resp_maps import ResponsibilityMaps
+from rex_xai.prediction import Prediction
 
 
 def subbox(tree, name, max_depth, min_size, mode, r_map=None):
